API is working with all verbs.

+++ Backend_.Net_3_1/LaunchTimeDB.Application/InputModels/Facilitators/FacilitatorInputModel.cs
@@ -4,8 +4,8 @@ namespace LaunchTimeDB.Application.InputModels.Facilitators
 {
     public class FacilitatorInputModel
     {
-        public int? Id { get; set; }
-        public int RestaurantId { get; set; }
+        public long? Id { get; set; }
+        public long RestaurantId { get; set; }
         public DateTime LaunchDate { get; set; }
     }
 }
